--comm

+++ model/comment.go
@@ -5,8 +5,8 @@ import "time"
 type CommentStore interface {
 	GetAllBySlug(string) ([]*Comment, error)
 	GetByID(uint64) (*Comment, error)
-	Create(*Comment) (int, error)
-	Delete(*Comment) (int, error)
+	Create(*Comment) (int64, error)
+	Delete(*Comment) (int64, error)
 }
 
 type Comment struct {
